+++ src/Forms/Fields/Concerns/HasValue.php
@@ -53,7 +53,7 @@ trait HasValue
 
         // Relationships are retrieved as collections - if they
         // are empty, we can return the default instead
-        if(is_countable($this->value) && count($this->value) == 0) {
+        if (is_countable($this->value) && count($this->value) == 0) {
             $default = $this->getDefault($locale);
 
             return (is_countable($default) && count($default) > 0) ? $default : $this->value;
@@ -82,11 +82,11 @@ trait HasValue
             }
 
             // Dotted syntax as support for array casts. We can fetch nested values as: days.0.am, days.0.pm, ...
-            if(str_contains($this->getColumnName(), '.')) {
+            if (str_contains($this->getColumnName(), '.')) {
                 $column = substr($this->getColumnName(), 0, strpos($this->getColumnName(), '.'));
                 $key = substr($this->getColumnName(), strpos($this->getColumnName(), '.') + 1);
 
-                if(is_array($model->{$column})) {
+                if (is_array($model->{$column})) {
                     return data_get($model->{$column}, trim($key, '.'));
                 }
             }
@@ -95,7 +95,7 @@ trait HasValue
 
             // Relationships are retrieved as collections - if they
             // are empty, we can return the default instead
-            if(is_countable($value) && count($value) == 0) {
+            if (is_countable($value) && count($value) == 0) {
                 $default = $this->getDefault($locale);
 
                 return (is_countable($default) && count($default) > 0) ? $default : $value;