+++ tests/Unit/UnitTestHelpers.php
@@ -13,4 +13,13 @@ trait UnitTestHelpers
 
         return $method->invokeArgs($object, $parameters);
     }
+
+    protected function getPrivateProperty($object, $propertyName)
+    {
+        $reflection = new \ReflectionClass(get_class($object));
+        $property = $reflection->getProperty($propertyName);
+        $property->setAccessible(true);
+
+        return $property->getValue($object);
+    }
 }