+++ CHANGELOG.md
@@ -6,7 +6,31 @@ principles.
 
 ## Unreleased
 
+## 0.8.23 - 2024-09-10
+
+-   Changed: Exports are now stored in the `storage/app/exports` folder.
+-   Changed: squanto is now using the project editor styling and scripts.
+-   Fixed: isolate all throwables while exporting all resources.
+-   Fixed: When no nav items, the title of empty nav block was still shown.
+
+## 0.8.22 - 2024-07-30
+
+-   Fixed: Issue where asset field values weren't saved on asset creation.
+
+## 0.8.21 - 2024-07-29
+
+-   Fixed: Import issue where text import of non-string values, like numbers would break squanto translations.
+-   Fixed: Import issue where menu headers are not labeled as expected. Now throws an error on import.
+
+## 0.8.20 - 2024-07-22
+
+-   Fixed: Could not save slug containing slashes on a nestable page.
+-   Added: Export plugin to export and import text translations for page resources, menu and static squanto content.
+
+## 0.8.19 - 2024-06-06
+
 -   Upgraded `thinktomorrow\url` dependency to `3.0.0`.
+-   Fixed: issue where repeat field would not work on create of fragment with custom dependencies.
 -   Fixed: multiple select filter options were not all used as filter in the query, only the first one.
 -   Fixed: Editor scripts were not loaded on the create page. Now the Html field works as intended on create.
 -   Added: `src/Forms/Layouts/Card` HasCollapsible trait to toggle content visibility.